migrated pip install to checkSys to avoid missing deps
