+++ environmentMngr.py
@@ -243,27 +243,5 @@ def downloadFile(fileName, downloadURL):
                 out.write(data)
         r.release_conn()
 
-def installPipPackage():
-    from sys import platform
-    toolName = "pip3"
-    downloadUrl = "https://bootstrap.pypa.io/get-pip.py"
-    fileName = "get-pip.py"
-    
-    if platform == "linux" or platform == "linux2" or platform == "linux-gnu":
-        if not checkToolLinux(toolName):
-            checkAndUpgradeOSPackageVersions('python3-pip') # Install PIP3
-        # If package manager fails to install, try using the bootstrap script
-        if not checkToolLinux(toolName):
-            downloadFile(fileName, downloadUrl)
-            os.system('python3 get-pip.py') # Install PIP3
 
-    elif platform == "darwin":
-        if not checkToolLinux(toolName):
-            downloadFile(fileName, downloadUrl)
-            os.system('python get-pip.py') # Install PIP3
-
-    elif platform == "win32" or platform == "win64":
-        if not checkToolWindows(toolName):
-            downloadFile(fileName, downloadUrl)
-            os.system('py get-pip.py') # Install PIP3
